% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustq.R
\name{adjustq_load_example}
\alias{adjustq_load_example}
\title{Load AdjustQ example}
\usage{
adjustq_load_example(sim = TRUE, ...)
}
\arguments{
\item{sim}{Logical; include simulation in AdjustQ. Default TRUE.}

\item{...}{Additional arguments passed to \code{\link[=adjustq]{adjustq()}}}
}
\value{
A data.frame (see \code{\link[=adjustq]{adjustq()}})
}
\description{
Runs AdjustQ using bundled NRKW1 data.
}
