% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_elev_curve}
\alias{area_elev_curve}
\title{Area-elevation curve for the zones of SFLN2}
\format{
A data.frame with 21 rows and 3 columns:
\describe{
\item{quantile}{Cumulative area fraction of the basin below the reference elevation}
\item{SFLN2-1}{Reference elevations for zone 1 (ft)}
\item{SFLN2-2}{Reference elevations for zone 2 (ft)}
}
}
\usage{
area_elev_curve
}
\description{
A dataset containing the percent area covered at a complete range of
elevations for the two zones of Salmon Falls Creek (SFLN2). Used as
input to \code{\link[=rsnwelev]{rsnwelev()}}.
}
\keyword{datasets}
