% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrkw1_daily_flow}
\alias{nrkw1_daily_flow}
\title{NRKW1 observed daily flow}
\format{
A data.frame with columns:
\describe{
\item{year,month,day}{Date columns}
\item{flow_cfs}{Daily average streamflow (cfs)}
\item{Source}{Data source identifier}
}
}
\usage{
nrkw1_daily_flow
}
\description{
Daily average observed streamflow at NRKW1.
}
\keyword{datasets}
