% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrkw1_forcing}
\alias{nrkw1_forcing}
\title{NRKW1 forcing data}
\format{
A named list of 2 data.frames (NRKW1-1, NRKW1-2), each with columns:
\describe{
\item{year,month,day,hour}{Datetime columns}
\item{map_mm}{Mean areal precipitation (mm)}
\item{mat_degc}{Mean areal temperature (deg C)}
\item{ptps}{Fraction of precipitation as snow}
}
}
\usage{
nrkw1_forcing
}
\description{
6-hour forcing data for Nooksack River at North Cedarville (NRKW1),
2 zones, period of record.
}
\keyword{datasets}
