% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrkw1_inst_flow}
\alias{nrkw1_inst_flow}
\title{NRKW1 observed instantaneous flow}
\format{
A data.frame with columns:
\describe{
\item{year,month,day,hour}{Datetime columns}
\item{flow_cfs}{Instantaneous streamflow (cfs)}
\item{Source}{Data source identifier}
}
}
\usage{
nrkw1_inst_flow
}
\description{
Instantaneous (6-hour) observed streamflow at NRKW1.
}
\keyword{datasets}
