% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrkw1_pars}
\alias{nrkw1_pars}
\title{NRKW1 optimal parameters}
\format{
A data.frame with 5 columns:
\describe{
\item{p_name}{Unique parameter identifier}
\item{name}{Parameter name}
\item{type}{Model type (sac, snow, uh, fa, lagk)}
\item{zone}{Zone or upstream reach name}
\item{value}{Parameter value}
}
}
\usage{
nrkw1_pars
}
\description{
Optimal parameters from NWRFC autocalibration for Nooksack River at
North Cedarville, WA (USGS-12210700). 2-zone SAC-SMA/SNOW17/UH model
with 3-reach Lag-K routing (6 hour timestep).
}
\keyword{datasets}
