% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrkw1_upflow}
\alias{nrkw1_upflow}
\title{NRKW1 upstream flow data}
\format{
A named list of 3 data.frames, each with columns:
\describe{
\item{year,month,day,hour}{Datetime columns}
\item{flow_cfs}{Streamflow (cfs)}
}
}
\usage{
nrkw1_upflow
}
\description{
6-hour upstream flow timeseries for 3 upstream reaches (MFNW1, NFNW1, NSSW1)
routed via Lag-K to Nooksack River at North Cedarville.
}
\keyword{datasets}
