% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwsrfs-run.R
\name{print.nwsrfs_run}
\alias{print.nwsrfs_run}
\title{Print method for nwsrfs_run objects}
\usage{
\method{print}{nwsrfs_run}(x, ...)
}
\arguments{
\item{x}{An nwsrfs_run object}

\item{...}{Ignored}
}
\value{
x (invisibly)
}
\description{
Print method for nwsrfs_run objects
}
