% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sac_snow}
\alias{sac_snow}
\title{Execute SAC-SMA, SNOW17, return total channel inflow per zone, and model states}
\usage{
sac_snow(dt_hours, forcing, pars, return_states = FALSE)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{forcing}{data frame with with columns for forcing inputs}

\item{pars}{sac parameters}

\item{return_states}{logical value indicating if the states should be output as well as the tci}
}
\value{
data.frame (1 column per zone) of unrouted channel inflow (tci), sac states
uztwc, uzfwc, lztwc, lzfsc, lzfpc, adimc, and snow water equivalent (swe),
and adjusted forcing data.
}
\description{
Execute SAC-SMA, SNOW17, return total channel inflow per zone, and model states
}
\examples{
\dontrun{
data(nrkw1_forcing)
data(nrkw1_pars)
dt_hours <- 6
flow <- sac_snow(dt_hours, nrkw1_forcing, nrkw1_pars)
}
}
