% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sac_snow_uh}
\alias{sac_snow_uh}
\title{Execute SAC-SMA, SNOW17 and UH with given parameters}
\usage{
sac_snow_uh(dt_hours, forcing, pars)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{forcing}{data frame with columns for forcing inputs}

\item{pars}{sac parameters}
}
\value{
Vector of routed flow in cfs
}
\description{
Execute SAC-SMA, SNOW17 and UH with given parameters
}
\examples{
\dontrun{
data(nrkw1_forcing)
data(nrkw1_pars)
dt_hours <- 6
flow_cfs <- sac_snow_uh(dt_hours, nrkw1_forcing, nrkw1_pars)
}
}
