% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfln2_daily_flow}
\alias{sfln2_daily_flow}
\title{SFLN2 observed daily flow}
\format{
A data.frame with columns:
\describe{
\item{year,month,day}{Date columns}
\item{flow_cfs}{Daily average streamflow (cfs)}
\item{Source}{Data source identifier}
}
}
\usage{
sfln2_daily_flow
}
\description{
Daily average observed streamflow at SFLN2.
}
\keyword{datasets}
