% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfln2_forcing}
\alias{sfln2_forcing}
\title{SFLN2 forcing data}
\format{
A named list of 3 data.frames (SFLN2-1, SFLN2-2, SFLN2-CU),
each with columns:
\describe{
\item{year,month,day,hour}{Datetime columns}
\item{map_mm}{Mean areal precipitation (mm)}
\item{mat_degc}{Mean areal temperature (deg C)}
\item{ptps}{Fraction of precipitation as snow}
}
}
\usage{
sfln2_forcing
}
\description{
6-hour forcing data for Salmon Falls Creek (SFLN2), 2 zones plus
a consumptive use zone, period of record.
}
\keyword{datasets}
