% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfln2_inst_flow}
\alias{sfln2_inst_flow}
\title{SFLN2 observed instantaneous flow}
\format{
A data.frame with columns:
\describe{
\item{year,month,day,hour}{Datetime columns}
\item{flow_cfs}{Instantaneous streamflow (cfs)}
\item{Source}{Data source identifier}
}
}
\usage{
sfln2_inst_flow
}
\description{
Instantaneous (6-hour) observed streamflow at SFLN2.
}
\keyword{datasets}
