% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfln2_pars}
\alias{sfln2_pars}
\title{SFLN2 optimal parameters}
\format{
A data.frame with 5 columns:
\describe{
\item{p_name}{Unique parameter identifier}
\item{name}{Parameter name}
\item{type}{Model type (sac, snow, uh, fa, chanloss, consuse)}
\item{zone}{Zone name}
\item{value}{Parameter value}
}
}
\usage{
sfln2_pars
}
\description{
Optimal parameters from NWRFC autocalibration for Salmon Falls Creek
NR San Jacinto NV (USGS-13105000). 2-zone SAC-SMA/SNOW17/UH model
with chanloss and consuse (6 hour timestep).
}
\keyword{datasets}
