% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwsrfs-run.R
\name{update_pars}
\alias{update_pars}
\title{Update parameters and re-run model chain}
\usage{
update_pars(run, new_pars)
}
\arguments{
\item{run}{An "nwsrfs_run" object from \code{\link[=nwsrfs_run]{nwsrfs_run()}} or \code{\link[=load_example]{load_example()}}}

\item{new_pars}{A data.frame with columns "p_name" and "value" containing
parameters to update. All p_name values must exist in the current pars.}
}
\value{
A new "nwsrfs_run" object with updated parameters
}
\description{
Update parameters and re-run model chain
}
