% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{fa_adj_nwrfc}
\alias{fa_adj_nwrfc}
\title{Conduct NWRFC style forcing adjustments}
\usage{
fa_adj_nwrfc(
  dt_hours,
  forcing,
  pars,
  climo = NULL,
  dry_run = FALSE,
  return_climo = FALSE
)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{forcing}{data frame with with columns for forcing inputs}

\item{pars}{sac parameters}

\item{climo}{climotology matrix}

\item{dry_run}{Do a run without any forcing adjustments, only compute pet and etd}

\item{return_climo}{Return the computed climo, instead of adjustments}
}
\value{
Matrix (1 column per zone) of unrouted channel inflow
}
\description{
Conduct NWRFC style forcing adjustments
}
\examples{
\dontrun{
data(nrkw1_forcing)
data(nrkw1_pars)
dt_hours <- 6
adj <- fa_adj_nwrfc(dt_hours, nrkw1_forcing, nrkw1_pars)
}
}
