% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwsrfsr.R
\docType{package}
\name{nwsrfsr}
\alias{nwsrfsr-package}
\alias{nwsrfsr}
\title{nwsrfsr: NWS River Forecast System Hydrologic Models}
\description{
R interface to NWSRFS Fortran hydrologic models used operationally by NOAA's
National Weather Service River Forecast Centers. Includes SAC-SMA soil moisture
accounting, SNOW-17 snow accumulation/ablation, gamma unit hydrograph routing,
Lag-K channel routing, channel loss/gain, and consumptive use modules.
}
\section{Low-level model wrappers}{

Individual model components callable via Fortran interface:
\code{\link[=sac_snow]{sac_snow()}}, \code{\link[=uh]{uh()}}, \code{\link[=lagk]{lagk()}}, \code{\link[=chanloss]{chanloss()}}, \code{\link[=consuse]{consuse()}}, \code{\link[=fa_nwrfc]{fa_nwrfc()}}
}

\section{High-level orchestration}{

\code{\link[=nwsrfs_run]{nwsrfs_run()}} reads NWRFC autocalibration directories and runs the full
model chain. \code{\link[=load_example]{load_example()}} provides bundled example data for NRKW1 and SFLN2.
}

\section{AdjustQ preprocessing}{

\code{\link[=adjustq]{adjustq()}} replicates the CHPS FEWS AdjustQ tool for creating upstream
flow timeseries from observed and simulated data.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NOAA-NWRFC/nwsrfs-hydro-models}
  \item \url{https://noaa-nwrfc.github.io/nwsrfs-hydro-models/}
  \item Report bugs at \url{https://github.com/NOAA-NWRFC/nwsrfs-hydro-models/issues}
}

}
\author{
\strong{Maintainer}: Cameron Bracken \email{cameron.bracken@pnnl.gov} (\href{https://orcid.org/0000-0003-1917-402X}{ORCID})

Authors:
\itemize{
  \item Geoffrey Walters \email{geoffrey.walters@noaa.gov} (\href{https://orcid.org/0009-0009-3804-8898}{ORCID})
}

Other contributors:
\itemize{
  \item Eric Anderson (Original SNOW-17 and SAC-SMA Fortran code (NOAA/HRL)) [contributor]
  \item George F. Smith (Original LAG-K Fortran code (NOAA/HRL)) [contributor]
  \item Janice M. Lewis (Interpolation and LAG-K Fortran code (NOAA/HRL)) [contributor]
  \item John E. Pask (Original sorting routines (LLNL); see inst/COPYRIGHTS) [contributor]
  \item Ondrej Certik (MIT-licensed fortran-utils sorting and types modules; see inst/COPYRIGHTS) [contributor]
  \item John Burkardt (MIT-licensed implementation of Brent's zero finder; see inst/COPYRIGHTS) [contributor]
  \item National Oceanic and Atmospheric Administration (International-rights holder for U.S. Government-authored components, including legacy NWSRFS Fortran and contributions by NOAA employees (C. Bracken through 2022-06-30, G. Walters); see inst/COPYRIGHTS) [copyright holder]
  \item Battelle Memorial Institute (Copyright holder for PNNL-authored contributions by C. Bracken from 2022-07-01 onward; see inst/COPYRIGHTS) [copyright holder]
}

}
