% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{rsnwelev}
\alias{rsnwelev}
\title{Replace ptps column with ptps derived using rain snow line code (lapse rate + MAT)}
\usage{
rsnwelev(forcing, pars, ae_tbl)
}
\arguments{
\item{forcing}{data frame with columns for forcing inputs}

\item{pars}{rsnwelev and snow17 parameters}

\item{ae_tbl}{data.table with col1 containing quantile info and subsequent col with elev for each zone}
}
\value{
Matrix (1 column per zone) of the forcing input argument with ptps replaced
with that derived from rsnwelev model
}
\description{
Replace ptps column with ptps derived using rain snow line code (lapse rate + MAT)
}
\examples{
\dontrun{
data(nrkw1_forcing)
data(nrkw1_pars)
forcing_adj <- rsnwelev(nrkw1_forcing, nrkw1_pars, area_elev_curve)
}
}
